29-poisonous-mushrooms/requirements.txt
cycler==0.10.0
joblib==0.15.1
kiwisolver==1.2.0
llvmlite==0.30.0
matplotlib==3.2.1
numba==0.46.0
numpy==1.18.4
pandas==1.0.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
threadpoolctl==2.0.0
umap-learn==0.3.9
